refactor: 타입 따로 관리
