+++ src/components/FilterTabButton/FilterTabButton.tsx
@@ -1,17 +1,9 @@
 'use client'
 import { useState, useEffect } from 'react'
 import * as styles from './FilterTabButton.css'
-interface FilterTabButtonProps {
-  text: string
-  onSelect: (text: string) => void // 상위 컴포넌트로 전달할 함수
-  selectedTab: string
-}
+import { filterTabButton } from '@/types/filterTabButton'
 
-const FilterTabButton = ({
-  text,
-  onSelect,
-  selectedTab,
-}: FilterTabButtonProps) => {
+const FilterTabButton = ({ text, onSelect, selectedTab }: filterTabButton) => {
   const [isSelected, setIsSelected] = useState<boolean>(false)
 
   // 버튼이 마운트 될 때 선택되었는지 확인
